requirements.txt
azure-ai-textanalytics==5.3.0b2
PyPDF2
python-dotenv
tfidf
scikit-learn
azure-cognitiveservices-speech
pydub
